chunk: split vec_chunk into its own file; require Chunk::TARGET
Move the worked `vec_chunk` reference impl (and its tests) out of the
1.5k-line `chunk.rs` into `chunk/vec_chunk.rs`, leaving `chunk/mod.rs`
holding just the `Chunk` trait and the generic harness. This makes room
for sibling impls (e.g. a columnar `col_chunk`) without a single massive
file; the module's own docs already frame `vec_chunk` as the reference
shape a next implementor follows.

Also drop the `= 1024` default on `Chunk::TARGET`, making it a required
associated const. The right value is layout-dependent (it trades against
the row-major vs. columnar crossover), and the sole impl already set it
explicitly, so the default had no users and only risked a silent,
mistuned inheritance for the next implementor.
